Queries: fix: do not allow to overwrite an alias by mistake
Setting the same alias on a different device should not be allowed.
Before this change this operation would cause the previous device reporting an
alias that doesn't own on its device status.
Beware: this change also forces a different behavior: an alias must be deleted
before to be reasigned to a different device.